requirements.txt
Django~=3.2.10
Pillow
djangorestframework
mysqlclient
channels
daphne